+++ boot/index.ts
@@ -7,6 +7,7 @@ import { CoursePaymentStorage } from "@/services/course-payments-storage";
 import { ContactStorage } from "@/services/contacts-storage";
 import { PasswordResetAttemptStorage } from "@/services/password-reset-attempts-storage";
 import { RabbiBioStorage } from "@/services/rabbi-bio-storage";
+import { CommunityUserStorage } from "@/services/community-users-storage";
 
 export async function Boot() {
   await UserStorage.ensureIndexes();
@@ -17,5 +18,6 @@ export async function Boot() {
   await ContactStorage.ensureIndexes();
   await PasswordResetAttemptStorage.ensureIndexes();
   await RabbiBioStorage.ensureIndexes();
+  await CommunityUserStorage.ensureIndexes();
   await OperatorStorage.ensureDefaultOperator();
 }